inline more of the html straight into the route
