--- www/html/templates.ts
@@ -1,5 +1,0 @@
-import useAppHtml from "./app.html.tsx";
-import IndexHtml from "./index.html.tsx";
-import DocumentHtml from "./document.html.tsx";
-
-export { DocumentHtml, IndexHtml, useAppHtml };
